% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{aprob_presidencial_wide2}
\alias{aprob_presidencial_wide2}
\title{Aprobación presidencial (wide 2)}
\description{
Aprobación presidencial por trimestre, 2000-2014, compilada por Catherine Reyes-Housholder. Se suman algunas variables del proyecto Quality of Government (2017). Formato wide (2).
}
\references{
Reyes-Housholder, C. (Forthcoming). A Theory of Gender’s Role on Presidential Approval Ratings in Corrupt Times. Political Research Quarterly.

\url{https://qog.pol.gu.se/data}
}
\keyword{data}
