% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{aprobacion_wide1}
\alias{aprobacion_wide1}
\title{Aprobación presidencial (wide)}
\description{
Datos de aprobación presidencial por trimestre, 2000-2014, compilada por Catherine Reyes-Housholder. Se suman algunas variables del proyecto Quality of Government (2017). Formato wide.
}
\references{
Reyes-Housholder, C. (2019). A Theory of Gender’s Role on Presidential Approval Ratings in Corrupt Times. Political Research Quarterly.

\url{https://qog.pol.gu.se/data}
}
\keyword{data}
