% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{bienestar}
\alias{bienestar}
\title{Bienestar en Latinoamérica}
\description{
Datos de bienestar en Latinoamérica, por Huber & Stephens (2012)
}
\references{
\url{http://huberandstephens.web.unc.edu/common-works/data/}
}
\keyword{data}
