% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{bienestar_la}
\alias{bienestar_la}
\title{Bienestar en Latinoamérica}
\description{
Bienestar en Latinoamérica
}
\references{
\url{http://huberandstephens.web.unc.edu/common-works/data/}
}
\keyword{data}
