% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{cinc_cow}
\alias{cinc_cow}
\title{Índice de capacidades materiales nacionales, de Correlates of War (CoW). Versión 5}
\description{
Índice de capacidades materiales nacionales, de Correlates of War (CoW). Versión 5
}
\references{
Singer, J. David, Stuart Bremer, and John Stuckey. (1972). "Capability Distribution, Uncertainty, and Major Power War, 1820-1965." in Bruce Russett (ed) Peace, War, and Numbers, Beverly Hills: Sage, 19-48.
}
\keyword{data}
