% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_only_chr.R
\name{cols_only_chr}
\alias{cols_only_chr}
\title{Indicates columns to load in readr::read_*, from a character vector.}
\usage{
cols_only_chr(x)
}
\arguments{
\item{x}{Character vector of column names}
}
\value{
A call to readr::cols_only()
}
\description{
This function indicates columns to load in readr::read_*.
}
\examples{
read_csv(readr_example("mtcars.csv"), col_types = cols_only_chr(c("mpg", "cyl")))
}
