% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{copatrocinio_arg}
\alias{copatrocinio_arg}
\title{Copatrocinio en el Senado argentino (1983)}
\description{
Datos de presentación conjunta de proyectos de ley en el Senado argentino
}
\references{
Alemán, E., Calvo, E., Jones, M. P., & Kaplan, N. (2009). Comparing Cosponsorship and Roll‐Call Ideal Points. Legislative Studies Quarterly, 34(1), 87-116.
}
\keyword{data}
