% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{crimen_brasil}
\alias{crimen_brasil}
\title{Crimen en Brasil}
\description{
Base de datos de crimen en los estados de Brasil.
}
\references{
Freire, D. (2018). Evaluating the effect of homicide prevention strategies in Sao Paulo, Brazil: A synthetic control approach. Latin American Research Review, 53(2).
}
\keyword{data}
