% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{datos_municipales}
\alias{datos_municipales}
\title{Datos de municipios chilenos (1992-2012)}
\description{
Base de datos obtenida por Soledad Araya, con variables para municipios chilenos entre 1992 y 2012.
}
\keyword{data}
