% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{dem_directa}
\alias{dem_directa}
\title{Democracia directa}
\description{
Democracia directa
}
\references{
Altman, D. (2018). Citizenship and Contemporary Direct Democracy. Cambridge University Press.
}
\keyword{data}
