% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{democracia_directa}
\alias{democracia_directa}
\title{Democracia directa}
\description{
Datos de instituciones de democracia directa en el mundo, compilados por David Altman (2018).
}
\references{
Altman, D. (2018). Citizenship and Contemporary Direct Democracy. Cambridge University Press.
}
\keyword{data}
