% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{discursos_allende}
\alias{discursos_allende}
\title{Discursos de Allende}
\description{
discursos públicos de Salvador Allende entre finales de 1969 y el 11 de septiembre de 1973.
}
\references{
\url{https://www.marxists.org/espanol/allende/index.htm}
}
\keyword{data}
