% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{edad_educ}
\alias{edad_educ}
\title{Edad y educación como confounders}
\description{
Datos simulados, en los que la edad y la educación son confounders para el tratamiento de usar redes de mosquitos. Creados por Andrew Heiss.

Datos de los estados de Brasil, compilados por Freire (2018)
}
\references{
Freire, D. (2018). Evaluating the effect of homicide prevention strategies in são paulo, brazil: A synthetic control approach. Latin American Research Review, 53(2), 231. \url{https://doi.org/10.25222/larr.334}
}
\keyword{data}
