% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{eeuu_brasil}
\alias{eeuu_brasil}
\title{Proximidad política entre EEUU y Brasil}
\description{
Datos de proximidad política entre EEUU y Brasil, generados a partir de sus votaciones en la Asamblea General de las Naciones Unidas. Datos originales de Voeten (2012).
}
\references{
Voeten, E. (2012). Data and analyses of voting in the un general assembly (SSRN Scholarly Paper ID 2111149). Social Science Research Network. \url{https://papers.ssrn.com/abstract=2111149}
}
\keyword{data}
