% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{lapop}
\alias{lapop}
\title{Encuesta LAPOP}
\description{
Encuesta LAPOP
}
\references{
Altman, D. (2018). Citizenship and Contemporary Direct Democracy. Cambridge University Press.
}
\keyword{data}
