% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_pvalues.R
\name{odds_pvalues}
\alias{odds_pvalues}
\title{p-values of exponentiated glm coefficients.}
\usage{
odds_pvalues(model)
}
\arguments{
\item{model}{glm model.}
}
\value{
A vector
}
\description{
This function generates the p-values of exponentiated glm coefficients.
The resulting vector is ready for texreg's "override.pvalues =" argument.
}
\details{
Partially adapted from a blog post by Andrew Heiss (2016).
}
\examples{
example_model <- glm(am ~ mpg, family = binomial(), data = mtcars)
texreg::screenreg(example_model,
                  override.coef    = exp(coef(example_model)),
                  override.se      = odds_se(example_model),
                  override.pvalues = odds_pvalues(example_model))
}
\references{
\url{https://www.andrewheiss.com/blog/2016/04/25/convert-logistic-regression-standard-errors-to-odds-ratios-with-r/}
}
