% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{pib_pc_america}
\alias{pib_pc_america}
\title{PIB per capita PPP en América}
\description{
Datos de PIB per capita con paridad de poder adquisitivo para los países de América, 1996-1998. Datos de los World Development Indicators del Banco Mundial.
}
\references{
\url{https://data.worldbank.org/indicator/NY.GDP.PCAP.PP.CD}
}
\keyword{data}
