% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{poltweets}
\alias{poltweets}
\title{Tweets políticos en Chile}
\description{
Tweets de diputados(as) chilenos(as), entre abril y junio de 2018. Compilados por Sebastián Huneeus.
}
\references{
\url{https://www.vanderbilt.edu/lapop/ab2016/AB2017-v18.0-Spa-170202_W.pdf}
}
\keyword{data}
