% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{qog}
\alias{qog}
\title{Datos de países del mundo, por Quality of Government (2017)}
\description{
Datos de países del mundo, por Quality of Government (2017)
}
\references{
\url{https://qog.pol.gu.se/data}
}
\keyword{data}
