% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{quiebre_democracia}
\alias{quiebre_democracia}
\title{Quiebres democráticos en Latinoamérica}
\description{
Datos de quiebres democráticos en América Latina, 1900-2010, por Mainwaring & Pérez-Liñán (2013).
}
\references{
Mainwaring, S., & Pérez-Liñán, A. (2013). Democracies and dictatorships in Latin America: emergence, survival, and fall. Cambridge University Press.
}
\keyword{data}
