% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{quiebres_dem_la}
\alias{quiebres_dem_la}
\title{Quiebres democráticos en Latinoamérica}
\description{
Quiebres democráticos en Latinoamérica
}
\references{
Mainwaring, S., & Pérez-Liñán, A. (2013). Democracies and dictatorships in Latin America: emergence, survival, and fall. Cambridge University Press.
}
\keyword{data}
