% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{red_mosquitos}
\alias{red_mosquitos}
\title{Redes de mosquitos}
\description{
Datos simulados, en los que se evalúa el tratamiento de redes de mosquitos en el riesgo de malaria. Creados por Andrew Heiss.
}
\keyword{data}
