% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{resumen_desordenado}
\alias{resumen_desordenado}
\title{Resumen desordenado del PIB per capita de países americanos}
\description{
Resumen "desordenado" del PIB per capita de países americanos, en el que los nombres de los países no siguen un estándar. Dato del PIB per capita desde los World Development Indicators del Banco Mundial.
}
\references{
\url{https://data.worldbank.org/indicator/NY.GDP.PCAP.PP.CD}
}
\keyword{data}
