% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{senadores_arg}
\alias{senadores_arg}
\title{Senadores de Argentina (1983)}
\description{
Datos de senadores argentinos(as), 1983.
}
\references{
Alemán, E., Calvo, E., Jones, M. P., & Kaplan, N. (2009). Comparing Cosponsorship and Roll‐Call Ideal Points. Legislative Studies Quarterly, 34(1), 87-116.
}
\keyword{data}
