% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{tratados}
\alias{tratados}
\title{Tratados internacionales}
\description{
Datos de adopción, ratificación, etc. de tratados internacionales. Compilados por Carsten Schulz en base al repositorio de las Naciones Unidas.
}
\references{
\url{https://www.carstenschulz.eu/}

\url{https://tratados.un.org/}
}
\keyword{data}
