% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{tratados_int}
\alias{tratados_int}
\title{Tratados internacionales}
\description{
Base de datos obtenida por Carsten Schulz.

Base de datos obtenida por Sebastián Huneeus, que contiene los tweets de los 155 diputados/as de Chile entre el 1 de mayo y el 10 de junio de 2018.
}
\keyword{data}
